before implementing form

+++ src/components/SnackItem/SnackItem.js
@@ -6,18 +6,34 @@ class Snack extends Component {
     return this.props.votes === undefined ? 0 : this.props.votes;
   }
 
+  voteForSnackHelper(snack, auth, voteForSnack){
+    if(auth){
+      return (
+        <button onClick={() => voteForSnack(snack)}>
+        vote
+        </button>
+      );
+    }
+  }
+
+  deleteSnackHelper(snack, auth, deleteSnack) {
+    if (auth) {
+      return (
+        <button onClick={() => deleteSnack(snack)}>
+        delete snack
+        </button>
+      )
+    }
+  }
+
   render() {
-    const { snack, voteForSnack, deleteSnack } = this.props;
+    const { snack, voteForSnack, deleteSnack, auth } = this.props;
     return (
       <li key={snack}>
         <span>{snack}</span>
-        <button onClick={() => voteForSnack(snack)}>
-          vote
-        </button>
-        <span>{this.voteCount}</span>
-        <button onClick={() => deleteSnack(snack)}>
-          delete snack
-        </button>
+        {this.voteForSnackHelper(snack, auth, voteForSnack)}
+        <span style={{marginLeft: "5px", marginRight: "5px"}}>{this.voteCount}</span>
+        {this.deleteSnackHelper(snack, auth, deleteSnack)}
       </li>
     );
   }
